fix: update data fetching from blockchain
